Options for 'vmcloak add' .vdi images can be stored in $name.opt files
This makes a fresh start easier, .vdi and .opt files can be dropped
into the data/vmcloak/image directory and they will be registered.

The /.vmcloak directory inside the container is ephemeral, apart from
the /.vmcloak/image directory which is bind mounted, this means all
snapshots will get lost when the container is destroyed.